endpoint: provide EndpointCreationManager via hive cell
Currently, the EndpointCreationManager that keeps track of the Endpoint
creation requests is created directly in the daemon.

In preparation for extracting the Endpoint API, this commit extracts
the `EndpointCreationManager` into its own Hive Cell `endpoint-api`
