+++ daemon/cmd/endpoint.go
@@ -28,11 +28,9 @@ import (
 	endpointmetadata "github.com/cilium/cilium/pkg/endpoint/metadata"
 	"github.com/cilium/cilium/pkg/endpoint/regeneration"
 	"github.com/cilium/cilium/pkg/ipam"
-	"github.com/cilium/cilium/pkg/k8s/client"
 	slim_corev1 "github.com/cilium/cilium/pkg/k8s/slim/k8s/api/core/v1"
 	"github.com/cilium/cilium/pkg/labels"
 	"github.com/cilium/cilium/pkg/labelsfilter"
-	"github.com/cilium/cilium/pkg/lock"
 	"github.com/cilium/cilium/pkg/logging/logfields"
 	"github.com/cilium/cilium/pkg/mac"
 	"github.com/cilium/cilium/pkg/option"
@@ -134,95 +132,6 @@ func (d *Daemon) errorDuringCreation(ep *endpoint.Endpoint, err error) (*endpoin
 	})
 	ep.Logger(daemonSubsys).Warn("Creation of endpoint failed", logfields.Error, err)
 	return nil, PutEndpointIDFailedCode, err
-}
-
-type endpointCreationRequest struct {
-	// cancel is the cancellation function that can be called to cancel
-	// this endpoint create request
-	cancel context.CancelFunc
-
-	// endpoint is the endpoint being added in the request
-	endpoint *endpoint.Endpoint
-
-	// started is the timestamp on when the processing has started
-	started time.Time
-}
-
-type endpointCreationManager struct {
-	mutex     lock.Mutex
-	clientset client.Clientset
-	requests  map[string]*endpointCreationRequest
-}
-
-func newEndpointCreationManager(cs client.Clientset) *endpointCreationManager {
-	return &endpointCreationManager{
-		requests:  map[string]*endpointCreationRequest{},
-		clientset: cs,
-	}
-}
-
-func (m *endpointCreationManager) NewCreateRequest(ep *endpoint.Endpoint, cancel context.CancelFunc) {
-	// Tracking is only performed if Kubernetes pod names are available.
-	// The endpoint create logic already ensures that IPs and CNI attachment ID
-	// are unique and thus tracking is not required outside of the
-	// Kubernetes context
-	if !ep.K8sNamespaceAndPodNameIsSet() || !m.clientset.IsEnabled() {
-		return
-	}
-
-	cepName := ep.GetK8sNamespaceAndCEPName()
-
-	m.mutex.Lock()
-	defer m.mutex.Unlock()
-
-	if req, ok := m.requests[cepName]; ok {
-		ep.Logger(daemonSubsys).Warn("Cancelling obsolete endpoint creating due to new create for same cep name")
-		req.cancel()
-	}
-
-	ep.Logger(daemonSubsys).Debug("New create request")
-	m.requests[cepName] = &endpointCreationRequest{
-		cancel:   cancel,
-		endpoint: ep,
-		started:  time.Now(),
-	}
-}
-
-func (m *endpointCreationManager) EndCreateRequest(ep *endpoint.Endpoint) bool {
-	if !ep.K8sNamespaceAndPodNameIsSet() || !m.clientset.IsEnabled() {
-		return false
-	}
-
-	cepName := ep.GetK8sNamespaceAndCEPName()
-
-	m.mutex.Lock()
-	defer m.mutex.Unlock()
-
-	if req, ok := m.requests[cepName]; ok {
-		if req.endpoint == ep {
-			ep.Logger(daemonSubsys).Debug("End of create request")
-			delete(m.requests, cepName)
-			return true
-		}
-	}
-
-	return false
-}
-
-func (m *endpointCreationManager) CancelCreateRequest(ep *endpoint.Endpoint) {
-	if m.EndCreateRequest(ep) {
-		ep.Logger(daemonSubsys).Warn("Cancelled endpoint create request due to receiving endpoint delete request")
-	}
-}
-
-func (m *endpointCreationManager) DebugStatus() (output string) {
-	m.mutex.Lock()
-	defer m.mutex.Unlock()
-
-	for _, req := range m.requests {
-		output += fmt.Sprintf("- %s: %s\n", req.started.String(), req.endpoint.String())
-	}
-	return
 }
 
 // createEndpoint attempts to create the endpoint corresponding to the change
